+++ HISTORY.rst
@@ -26,6 +26,7 @@ Latest (unreleased)
 * Remove signatures on leads
 * update transitions checks in crm
 * Add NDA Leads in customer's form
+* Update filters & rules for NDA & stage3
 
 **Bugfixes**